Reverted lost change of type for port variable.

+++ libteol0/teonet_l0_client.c
@@ -684,7 +684,7 @@ int teoLNullReadEventLoop(teoLNullConnectData *con, int timeout)
  * @retval teoLNullConnectData::status==-2 - HOST NOT FOUND error
  * @retval teoLNullConnectData::status==-3 - Client-connect() error
  */
-teoLNullConnectData* teoLNullConnectE(const char *server, int port, teoLNullEventsCb event_cb,
+teoLNullConnectData* teoLNullConnectE(const char *server, int16_t port, teoLNullEventsCb event_cb,
         void *user_data, PROTOCOL connection_flag)
 {
 
@@ -787,7 +787,7 @@ teoLNullConnectData* teoLNullConnectE(const char *server, int port, teoLNullEven
  * @retval teoLNullConnectData::status==-2 - HOST NOT FOUND error
  * @retval teoLNullConnectData::status==-3 - Client-connect() error
  */
-teoLNullConnectData* teoLNullConnect(const char *server, int port, PROTOCOL connection_flag)
+teoLNullConnectData* teoLNullConnect(const char *server, int16_t port, PROTOCOL connection_flag)
 {
     return teoLNullConnectE(server, port, NULL, NULL, connection_flag);
 }
